Improve array index parsing
- Return the new field name length to avoid a second find
- Better handling of invalid field names

+++ server/modules/selva/lib/util/cstrings.h
@@ -17,7 +17,7 @@ int stringlist_search(const char *list, const char *str);
 int stringlist_searchn(const char *list, const char *str, size_t n);
 size_t substring_count(const char *string, const char *substring, size_t n);
 
-int get_array_field_index(const char *field_str, size_t field_len, ssize_t *res);
+ssize_t get_array_field_index(const char *field_str, size_t field_len, ssize_t *res);
 
 /**
  * Calculate the number of instances of ch in s.
